Refine entity descriptions

+++ sql/setup_database.sql
@@ -14,7 +14,7 @@ name varchar(32) NOT NULL UNIQUE
 );
 
 CREATE TABLE tracked_activity (
-tracked_activity_id int8 primary key,
+tracked_activity_id SERIAL primary key,
 activity_blueprint_id integer references activity_blueprint(activity_blueprint_id) ON DELETE NO ACTION ON UPDATE NO ACTION,
 user_id varchar(64) references app_user(user_id) ON DELETE NO ACTION ON UPDATE NO ACTION,
 time_start timestamp NOT NULL,
